Move WCIF model files to own package
