--- webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/wcif/Room.kt
@@ -1,3 +1,0 @@
-package org.worldcubeassociation.tnoodle.server.webscrambles.wcif
-
-data class Room(val name: String, val activities: List<Activity>) : SafeNamed(name)
