+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/wcif/model/Room.kt
@@ -1,0 +1,6 @@
+package org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model
+
+import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.Activity
+import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.SafeNamed
+
+data class Room(val name: String, val activities: List<Activity>) : SafeNamed(name)
